Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 int main() {
     int a,b,c;
-    int d=(a+b+c)/3;
-    scanf("%d",&a);
-    scanf("%d",&b);
-    scanf("%d",&c);
-    printf("Average: %d",d);
+    float d=(a+b+c)/3.0;
+    scanf("%d", &a);
+    scanf("%d", &b);
+    scanf("%d", &c);
+    printf("Average: %.2f",d);
     return 0;
 }
